requirements.txt
# Data processing
pandas==2.0.3
numpy==1.24.3
scipy==1.11.3
pyarrow==12.0.0  # For parquet files

# Machine learning
scikit-learn==1.3.0
implicit==0.6.2
torch==2.0.1
lightfm==1.17

# API
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.4.2

# Utilities
tqdm==4.66.1
jupyter==1.0.0
matplotlib==3.7.3
seaborn==0.12.2
PyYAML==6.0.1
